refactor: util folder
