+++ README.md
@@ -57,7 +57,7 @@ PS.: It is important to add the `--template ui5-webcomponents-react-seed` at the
 
 - MockServer Engine (w/ `json-server`);
 
-- HTTP Request Engine (w/ `Request`, `APIProvider`, `BrowserProvider`);
+- API HTTP Request Engine (w/ `Request`, `APIProvider`, `BrowserProvider`);
 
 - Permission Engine (w/ `RouteValidator` and `ComponentValidator`).
 
@@ -74,7 +74,6 @@ PS.: It is important to add the `--template ui5-webcomponents-react-seed` at the
 Following one of the several recommendations for structuring files on a React based project, the seed use: [Grouping by file type](https://reactjs.org/docs/faq-structure.html#grouping-by-file-type).
 
 The only custom change we have incremented were the Custom Components and the folder for each project containing the `tests` artefacts.
-
 
 # Scripts Included
