+++ src/components/Shell/Shell.js
@@ -6,7 +6,7 @@ import { ShellBar } from '@ui5/webcomponents-react/lib/ShellBar';
 import { Avatar } from '@ui5/webcomponents-react/lib/Avatar';
 import { AvatarShape } from '@ui5/webcomponents-react/lib/AvatarShape';
 import { AvatarSize } from '@ui5/webcomponents-react/lib/AvatarSize';
-import BrowserProvider from '../../util/URL/Browser/BrowserProvider';
+import BrowserProvider from '../../util/browser/BrowserProvider';
 import PopoverListItems from '../Popover/List/PopoverListItems';
 
 const style = {
