+++ src/util/api/url/APIProvider.js
@@ -1,4 +1,4 @@
-import Provider from '../Provider';
+import URLProvider from '../../URLProvider';
 
 const URLs = {
   GET_USER_LOGGED: '/v1/user/logged',
@@ -8,6 +8,6 @@ const URLs = {
 
 export default {
   getUrl(key, replaceOptions) {
-    return Provider.getUrl(URLs, key, replaceOptions);
+    return URLProvider.getUrl(URLs, key, replaceOptions);
   },
 };
