+++ src/util/browser/BrowserProvider.test.js
